+++ CHANGELOG.md
@@ -11,11 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Added `compas_rhino.artists.BoxArtist.draw_collection`.
 * Added option to show/hide vertices and faces in `compas_rhino.artists.CapsuleArtist.draw`.
+* Added option to show/hide vertices and faces in `compas_rhino.artists.ConeArtist.draw`.
 
 ### Changed
 
-* Changed parameters of `compas_rhino.artists.BoxArtist.draw`.
+* Changed implementation of `compas_rhino.artists.BoxArtist.draw`.
 * Fixed bug in `compas.geometry.Capsule`.
+* Fixed bug in `compas.geometry.Cone`.
 
 ### Removed